# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamSync Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all servers
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a new server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a server by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    put:
      operationId: updateServer
      tags: [servers]
      summary: Update a server
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerBody"
      responses:
        "200":
          description: Server updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a server
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/test:
    post:
      operationId: testServerConnection
      tags: [servers]
      summary: Test connection to a server using its API key
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerTestResult"
        "404":
          description: Server not found
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Server:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [plex, jellyfin, emby]
        apiKey:
          type: string
        active:
          type: boolean
        userCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, url, type, apiKey, active, userCount, createdAt]
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [plex, jellyfin, emby]
        apiKey:
          type: string
        active:
          type: boolean
      required: [name, url, type, apiKey]
    UpdateServerBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [plex, jellyfin, emby]
        apiKey:
          type: string
        active:
          type: boolean
    ServerTestResult:
      type: object
      properties:
        success:
          type: boolean
        serverName:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        responseMs:
          type: integer
          nullable: true
        checks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              status:
                type: string
                enum: [ok, fail, skip]
              detail:
                type: string
                nullable: true
            required: [name, status]
        error:
          type: string
          nullable: true
      required: [success, checks]
...