# lib/api-zod/src/generated/types/createServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateServerBodyType } from "./createServerBodyType";

export interface CreateServerBody {
  name: string;
  url: string;
  type: CreateServerBodyType;
  apiKey: string;
  active?: boolean;
}