# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerType } from "./serverType";

export interface Server {
  id: number;
  name: string;
  url: string;
  type: ServerType;
  apiKey: string;
  active: boolean;
  userCount: number;
  createdAt: string;
}