# lib/api-zod/src/generated/types/serverTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerTestResultChecksItem } from "./serverTestResultChecksItem";

export interface ServerTestResult {
  success: boolean;
  serverName?: string | null;
  version?: string | null;
  responseMs?: number | null;
  checks: ServerTestResultChecksItem[];
  error?: string | null;
}