# lib/api-zod/src/generated/types/updateServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamSync Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateServerBodyType } from "./updateServerBodyType";

export interface UpdateServerBody {
  name?: string;
  url?: string;
  type?: UpdateServerBodyType;
  apiKey?: string;
  active?: boolean;
}